Генератор для целых значений

+++ Web/Web/Models/Settings/SettingsValues.cs
@@ -2,26 +2,26 @@
 {
     public class SettingsValues
     {
-        public decimal CountOfClasses { get; set; }
+        public int CountOfClasses { get; set; }
 
-        public decimal CountOfProperties { get; set; }
+        public int CountOfProperties { get; set; }
 
-        public decimal AwailablePropertyMin { get; set; }
+        public int AwailablePropertyMin { get; set; }
 
-        public decimal AwailablePropertyMax { get; set; }
+        public int AwailablePropertyMax { get; set; }
 
-        public decimal NormalPropertyMin { get; set; }
+        public int NormalPropertyMin { get; set; }
 
-        public decimal NormalPropertyMax { get; set; }
+        public int NormalPropertyMax { get; set; }
 
-        public decimal CountOfPeriods { get; set; }
+        public int CountOfPeriods { get; set; }
 
-        public decimal AwailablePeriodsMin { get => NormalPropertyMin; }
+        public int AwailablePeriodsMin { get => NormalPropertyMin; }
 
-        public decimal AwailablePeriodsMax { get => NormalPropertyMax; }
+        public int AwailablePeriodsMax { get => NormalPropertyMax; }
 
-        public decimal NormalPeriodsMin { get; set; }
+        public int NormalPeriodsMin { get; set; }
 
-        public decimal NormalPeriodsMax { get; set; }
+        public int NormalPeriodsMax { get; set; }
     }
 }
